refactor: add old price ticker placeholder (#706)
