+++ app/Http/Livewire/Navbar/PriceTicker.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Http\Livewire;
+namespace App\Http\Livewire\Navbar;
 
 use App\Facades\Network;
 use App\Facades\Settings;
@@ -11,22 +11,21 @@ use App\Services\Cache\NetworkStatusBlockCache;
 use App\Services\ExchangeRate;
 use App\Services\NumberFormatter;
 use Illuminate\Support\Str;
+use Illuminate\View\View;
 use Livewire\Component;
 
-/**
- * TODO: remove in a future update. Kept to prevent errors being spammed.
- *
- * @codeCoverageIgnore
- */
 final class PriceTicker extends Component
 {
     use HandlesSettings;
 
-    public string $price;
-
     public string $to;
 
     public bool $isAvailable = false;
+
+    /** @var mixed */
+    protected $listeners = [
+        'currencyChanged' => 'setValues',
+    ];
 
     public function mount(): void
     {
@@ -36,15 +35,17 @@ final class PriceTicker extends Component
     public function setValues(): void
     {
         $this->isAvailable = (new NetworkStatusBlockCache())->getIsAvailable(Network::currency(), Settings::currency());
-        $this->price       = $this->getPriceFormatted();
         $this->to          = Settings::currency();
 
         $this->dispatchBrowserEvent('has-loaded-price-data');
     }
 
-    public function render(): string
+    public function render(): View
     {
-        return '';
+        return view('livewire.navbar.price-ticker', [
+            'to'    => $this->to,
+            'price' => $this->getPriceFormatted(),
+        ]);
     }
 
     public function setCurrency(string $newCurrency): void
