+++ resources/views/livewire/navbar/price-ticker.blade.php
